Showcase Cut
The version of the game as of 5/15/2018. The game is completely furnished with visual and audio assets, and numerous bugs were worked out since previous versions. Unless any surprise bugs are discovered and fixed within the next two days, this will be the version of the game shown at the 2018 Student Showcase.